Moved the outlets and actions into a seperate file for the battery view
